# output/src/ApiApp/Controllers/EmployeesController.cs
// Generated by booster-apidev | 2026-04-06 19:37:30

using ApiApp.Data;
using ApiApp.DTOs.Employees;
using ApiApp.DTOs.Orders;
using ApiApp.Models;
using Microsoft.AspNetCore.Mvc;
using Microsoft.EntityFrameworkCore;

namespace ApiApp.Controllers;

[ApiController]
[Route("api/employees")]
public class EmployeesController : ControllerBase
{
    private readonly AppDbContext _db;

    public EmployeesController(AppDbContext db)
    {
        _db = db;
    }

    // GET /api/employees
    [HttpGet]
    public async Task<ActionResult<DTOs.PagedResult<EmployeeListItemDto>>> GetEmployees(
        [FromQuery] string? search,
        [FromQuery] int page = 1,
        [FromQuery] int pageSize = 20)
    {
        pageSize = Math.Min(pageSize, 100);
        page = Math.Max(page, 1);

        var query = _db.Employees.AsQueryable();

        if (!string.IsNullOrWhiteSpace(search))
        {
            var s = search.ToLower();
            query = query.Where(e =>
                (e.LastName != null && e.LastName.ToLower().Contains(s)) ||
                (e.FirstName != null && e.FirstName.ToLower().Contains(s)) ||
                (e.EmailAddress != null && e.EmailAddress.ToLower().Contains(s)));
        }

        var totalCount = await query.CountAsync();
        var items = await query
            .OrderBy(e => e.LastName).ThenBy(e => e.FirstName)
            .Skip((page - 1) * pageSize)
            .Take(pageSize)
            .Select(e => new EmployeeListItemDto
            {
                Id = e.Id,
                FirstName = e.FirstName,
                LastName = e.LastName,
                FullNameFNLN = (e.FirstName + " " + e.LastName).Trim(),
                FullNameLNFN = (e.LastName != null && e.FirstName != null)
                    ? e.LastName + ", " + e.FirstName
                    : (e.LastName ?? e.FirstName ?? ""),
                EmailAddress = e.EmailAddress,
                JobTitle = e.JobTitle,
                PrimaryPhone = e.PrimaryPhone,
                SecondaryPhone = e.SecondaryPhone,
                Title = e.Title
            })
            .ToListAsync();

        return Ok(new DTOs.PagedResult<EmployeeListItemDto>
        {
            Items = items,
            TotalCount = totalCount,
            Page = page,
            PageSize = pageSize
        });
    }

    // GET /api/employees/{id}
    [HttpGet("{id:int}")]
    public async Task<ActionResult<EmployeeDetailDto>> GetEmployee(int id)
    {
        var e = await _db.Employees.FindAsync(id);
        if (e == null) return NotFound(new { error = $"Employee {id} not found." });
        return Ok(MapToDetail(e));
    }

    // POST /api/employees
    [HttpPost]
    public async Task<ActionResult<EmployeeDetailDto>> CreateEmployee([FromBody] EmployeeCreateUpdateDto dto)
    {
        // Check unique WindowsUserName
        if (!string.IsNullOrWhiteSpace(dto.WindowsUserName))
        {
            var exists = await _db.Employees.AnyAsync(e => e.WindowsUserName == dto.WindowsUserName);
            if (exists)
                return BadRequest(new { errors = new { WindowsUserName = new[] { "WindowsUserName must be unique." } } });
        }

        var employee = new Employee
        {
            FirstName = dto.FirstName,
            LastName = dto.LastName,
            EmailAddress = dto.EmailAddress,
            JobTitle = dto.JobTitle,
            PrimaryPhone = dto.PrimaryPhone,
            SecondaryPhone = dto.SecondaryPhone,
            Title = dto.Title,
            Notes = dto.Notes,
            WindowsUserName = dto.WindowsUserName
        };

        _db.Employees.Add(employee);
        await _db.SaveChangesAsync();

        return CreatedAtAction(nameof(GetEmployee), new { id = employee.Id }, MapToDetail(employee));
    }

    // PUT /api/employees/{id}
    [HttpPut("{id:int}")]
    public async Task<ActionResult<EmployeeDetailDto>> UpdateEmployee(int id, [FromBody] EmployeeCreateUpdateDto dto)
    {
        var employee = await _db.Employees.FindAsync(id);
        if (employee == null) return NotFound(new { error = $"Employee {id} not found." });

        // Check unique WindowsUserName excluding self
        if (!string.IsNullOrWhiteSpace(dto.WindowsUserName))
        {
            var duplicateExists = await _db.Employees
                .AnyAsync(e => e.WindowsUserName == dto.WindowsUserName && e.Id != id);
            if (duplicateExists)
                return BadRequest(new { errors = new { WindowsUserName = new[] { "WindowsUserName must be unique." } } });
        }

        employee.FirstName = dto.FirstName;
        employee.LastName = dto.LastName;
        employee.EmailAddress = dto.EmailAddress;
        employee.JobTitle = dto.JobTitle;
        employee.PrimaryPhone = dto.PrimaryPhone;
        employee.SecondaryPhone = dto.SecondaryPhone;
        employee.Title = dto.Title;
        employee.Notes = dto.Notes;
        employee.WindowsUserName = dto.WindowsUserName;

        await _db.SaveChangesAsync();
        return Ok(MapToDetail(employee));
    }

    // DELETE /api/employees/{id}
    [HttpDelete("{id:int}")]
    public async Task<IActionResult> DeleteEmployee(int id)
    {
        var employee = await _db.Employees.FindAsync(id);
        if (employee == null) return NotFound(new { error = $"Employee {id} not found." });

        // Nullify FK on associated orders
        var orders = await _db.Orders.Where(o => o.EmployeeId == id).ToListAsync();
        foreach (var order in orders)
            order.EmployeeId = null;

        _db.Employees.Remove(employee);
        await _db.SaveChangesAsync();
        return NoContent();
    }

    // GET /api/employees/{id}/orders
    [HttpGet("{id:int}/orders")]
    public async Task<ActionResult<DTOs.PagedResult<OrderListItemDto>>> GetEmployeeOrders(
        int id,
        [FromQuery] int? top,
        [FromQuery] int page = 1,
        [FromQuery] int pageSize = 20)
    {
        var exists = await _db.Employees.AnyAsync(e => e.Id == id);
        if (!exists) return NotFound(new { error = $"Employee {id} not found." });

        var query = _db.Orders
            .Where(o => o.EmployeeId == id)
            .Include(o => o.Customer)
            .Include(o => o.Employee)
            .Include(o => o.OrderDetails)
            .OrderByDescending(o => o.OrderDate);

        if (top.HasValue && top.Value > 0)
        {
            var topItems = await query.Take(top.Value).ToListAsync();
            return Ok(new DTOs.PagedResult<OrderListItemDto>
            {
                Items = topItems.Select(MapOrderToListItem),
                TotalCount = topItems.Count,
                Page = 1,
                PageSize = top.Value
            });
        }

        pageSize = Math.Min(pageSize, 100);
        page = Math.Max(page, 1);
        var total = await query.CountAsync();
        var items = await query.Skip((page - 1) * pageSize).Take(pageSize).ToListAsync();

        return Ok(new DTOs.PagedResult<OrderListItemDto>
        {
            Items = items.Select(MapOrderToListItem),
            TotalCount = total,
            Page = page,
            PageSize = pageSize
        });
    }

    private static EmployeeDetailDto MapToDetail(Employee e) => new()
    {
        Id = e.Id,
        FirstName = e.FirstName,
        LastName = e.LastName,
        FullNameFNLN = e.FullNameFNLN,
        FullNameLNFN = e.FullNameLNFN,
        EmailAddress = e.EmailAddress,
        JobTitle = e.JobTitle,
        PrimaryPhone = e.PrimaryPhone,
        SecondaryPhone = e.SecondaryPhone,
        Title = e.Title,
        Notes = e.Notes,
        WindowsUserName = e.WindowsUserName,
        AddedBy = e.AddedBy,
        AddedOn = e.AddedOn,
        ModifiedBy = e.ModifiedBy,
        ModifiedOn = e.ModifiedOn
    };

    private static OrderListItemDto MapOrderToListItem(Order o) => new()
    {
        Id = o.Id,
        CustomerId = o.CustomerId,
        CustomerName = o.Customer?.CustomerName,
        EmployeeId = o.EmployeeId,
        EmployeeFullName = o.Employee != null
            ? $"{o.Employee.FirstName} {o.Employee.LastName}".Trim()
            : null,
        OrderDate = o.OrderDate,
        ShippedDate = o.ShippedDate,
        PaidDate = o.PaidDate,
        Status = o.Status.ToString(),
        StatusDisplay = o.Status.ToString(),
        OrderTotal = o.OrderTotal,
        ModifiedOn = o.ModifiedOn
    };
}