# output/src/ApiApp/Controllers/LookupController.cs
// Generated by booster-apidev | 2026-04-06 19:37:30

using ApiApp.Data;
using ApiApp.DTOs.Lookup;
using ApiApp.Models;
using Microsoft.AspNetCore.Mvc;
using Microsoft.EntityFrameworkCore;

namespace ApiApp.Controllers;

[ApiController]
[Route("api/lookup")]
public class LookupController : ControllerBase
{
    private readonly AppDbContext _db;

    public LookupController(AppDbContext db)
    {
        _db = db;
    }

    // GET /api/lookup/customers
    [HttpGet("customers")]
    public async Task<ActionResult<IEnumerable<CustomerLookupDto>>> GetCustomers()
    {
        var result = await _db.Customers
            .OrderBy(c => c.CustomerName)
            .Select(c => new CustomerLookupDto { Id = c.Id, CustomerName = c.CustomerName })
            .ToListAsync();

        return Ok(result);
    }

    // GET /api/lookup/employees
    [HttpGet("employees")]
    public async Task<ActionResult<IEnumerable<EmployeeLookupDto>>> GetEmployees()
    {
        var result = await _db.Employees
            .OrderBy(e => e.LastName).ThenBy(e => e.FirstName)
            .Select(e => new EmployeeLookupDto
            {
                Id = e.Id,
                FullNameFNLN = (e.FirstName + " " + e.LastName).Trim()
            })
            .ToListAsync();

        return Ok(result);
    }

    // GET /api/lookup/products
    [HttpGet("products")]
    public async Task<ActionResult<IEnumerable<ProductLookupDto>>> GetProducts()
    {
        var result = await _db.Products
            .OrderBy(p => p.ProductName)
            .Select(p => new ProductLookupDto
            {
                Id = p.Id,
                ProductCode = p.ProductCode,
                ProductName = p.ProductName,
                UnitPrice = p.UnitPrice
            })
            .ToListAsync();

        return Ok(result);
    }

    // GET /api/lookup/order-statuses
    [HttpGet("order-statuses")]
    public ActionResult<IEnumerable<OrderStatusLookupDto>> GetOrderStatuses()
    {
        var statuses = Enum.GetValues<OrderStatus>()
            .OrderBy(s => (int)s)
            .Select(s => new OrderStatusLookupDto
            {
                Value = s.ToString(),
                Display = s.ToString(),
                SortOrder = (int)s
            });

        return Ok(statuses);
    }
}